+++ ft8CN/app/src/main/java/com/bg7yoz/ft8cn/ft8transmit/OnDoTransmitted.java
@@ -1,9 +1,14 @@
 package com.bg7yoz.ft8cn.ft8transmit;
+/**
+ * 发射的回调
+ * @author BGY70Z
+ * @date 2023-03-20
+ */
 
 import com.bg7yoz.ft8cn.Ft8Message;
 
 public interface OnDoTransmitted {
     void onBeforeTransmit(Ft8Message message,int functionOder);
     void onAfterTransmit(Ft8Message message, int functionOder);
-    void onAfterGenerate(short[] data);
+    void onAfterGenerate(float[] data);
 }
